+++ jest.config.js
@@ -2,6 +2,7 @@
 module.exports = {
   preset: "ts-jest",
   setupFiles: ["fake-indexeddb/auto"],
+  snapshotSerializers: ["<rootDir>/src/__tests__/serializers/pathSerializer"],
   restoreMocks: true,
   testMatch: ["**/*.+(spec|test).ts?(x)"],
   testPathIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/examples/"],